+++ mpvpaper/Main.qml
@@ -1,4 +1,3 @@
-pragma ComponentBehavior: Bound
 import Qt.labs.folderlistmodel
 import QtQuick
 import Quickshell
@@ -6,6 +5,8 @@ import Quickshell.Io
 
 import qs.Commons
 import qs.Services.UI
+
+import "./main"
 
 Item {
     id: root
@@ -15,24 +16,71 @@ Item {
         pluginApi.pluginSettings.active || 
         false
 
-    readonly property string wallpapersFolder: 
-        pluginApi.pluginSettings.wallpapersFolder || 
-        pluginApi.manifest.metadata.defaultSettings.wallpapersFolder || 
-        "~/Pictures/Wallpapers"
+    readonly property bool automation:
+        pluginApi.pluginSettings.automation ||
+        false
+
+    readonly property string automationMode:
+        pluginApi.pluginSettings.automationMode ||
+        "random"
+
+    readonly property real automationTime:
+        pluginApi.pluginSettings.automationTime ||
+        5 * 60
 
     readonly property string currentWallpaper: 
         pluginApi.pluginSettings.currentWallpaper || 
         ""
+
+    readonly property bool hardwareAcceleration:
+        pluginApi.pluginSettings.hardwareAcceleration ||
+        false
+
+    readonly property bool isMuted:
+        pluginApi.pluginSettings.isMuted ||
+        false
+
+    readonly property bool isPlaying:
+        pluginApi.pluginSettings.isPlaying ||
+        false
 
     readonly property string mpvSocket: 
         pluginApi.pluginSettings.mpvSocket || 
         pluginApi.manifest.metadata.defaultSettings.mpvSocket || 
         "/tmp/mpv-socket"
 
-    // Thumbnail variables
-    property bool thumbCacheReady: false
-    property int _thumbGenIndex: 0
+    readonly property var oldWallpapers:
+        pluginApi.pluginSettings.oldWallpapers || 
+        ({})
 
+    readonly property string profile:
+        pluginApi.pluginSettings.profile ||
+        pluginApi.manifest.metadata.defaultSettings.profile ||
+        "default"
+
+    readonly property string fillMode:
+        pluginApi.pluginSettings.fillMode ||
+        pluginApi.manifest.metadata.defaultSettings.fillMode ||
+        "fit"
+    
+
+    readonly property bool thumbCacheReady:
+        pluginApi.pluginSettings.thumbCacheReady ||
+        false
+
+    readonly property real volume:
+        pluginApi.pluginSettings.volume ||
+        100
+
+    readonly property string wallpapersFolder: 
+        pluginApi.pluginSettings.wallpapersFolder || 
+        pluginApi.manifest.metadata.defaultSettings.wallpapersFolder || 
+        "~/Pictures/Wallpapers"
+
+
+    /***************************
+    * WALLPAPER FUNCTIONALITY
+    ***************************/
     function random() {
         if (wallpapersFolder === "" || folderModel.count === 0) {
             Logger.e("mpvpaper", "Empty wallpapers folder or no files found!");
@@ -48,6 +96,21 @@ Item {
         setWallpaper("");
     }
 
+    function nextWallpaper() {
+        if (wallpapersFolder === "" || folderModel.count === 0) {
+            Logger.e("mpvpaper", "Empty wallpapers folder or no files found!");
+            return;
+        }
+
+        Logger.d("mpvpaper", "Choosing next wallpaper...");
+
+        // Even if the file is not in wallpapers folder, aka -1, it sets the nextIndex to 0 then
+        const currentIndex = folderModel.indexOf(root.currentWallpaper);
+        const nextIndex = (currentIndex + 1) % folderModel.count;
+        const url = folderModel.get(nextIndex, "filePath");
+        setWallpaper(url);
+    }
+
     function setWallpaper(path) {
         if (root.pluginApi == null) {
             Logger.e("mpvpaper", "Can't set the wallpaper because pluginApi is null.");
@@ -58,133 +121,92 @@ Item {
         pluginApi.saveSettings();
     }
 
-    function setActive(isActive) {
-        if(root.pluginApi == null) {
-            Logger.e("mpvpaper", "Can't change active state because pluginApi is null.");
-            return;
-        }
 
-        pluginApi.pluginSettings.active = isActive;
-        pluginApi.saveSettings();
-    }
-
-    function thumbRegenerate() {
-        root.thumbCacheReady = false;
-        thumbProc.command = ["sh", "-c", `rm -rf ${thumbCacheFolder} && mkdir -p ${thumbCacheFolder}`]
-        thumbProc.running = true;
-    }
-
-    function thumbGeneration() {
-        while(root._thumbGenIndex < folderModel.count) {
-            const videoUrl = folderModel.get(root._thumbGenIndex, "fileUrl");
-            const thumbUrl = root.getThumbUrl(videoUrl);
-            root._thumbGenIndex++;
-            // Check if file already exists, otherwise create it with ffmpeg
-            if (thumbFolderModel.indexOf(thumbUrl) === -1) {
-                Logger.d("mpvpaper", `Creating thumbnail for video: ${videoUrl}`);
-
-                thumbProc.command = ["sh", "-c", `ffmpeg -y -i ${videoUrl} -vf "scale=1080:-1" -vframes:v 1 ${thumbUrl}`]
-                thumbProc.running = true;
-                return;
-            }
-        }
-
-        // The thumbnail generation has looped over every video and finished the generation.
-        root._thumbGenIndex = 0;
-        root.thumbCacheReady = true;
-    }
-
-    /* HELPER FUNCTIONALITY */
-    readonly property string thumbCacheFolder: ImageCacheService.wpThumbDir + "mpvpaper"
-
+    /***************************
+    * HELPER FUNCTIONALITY
+    ***************************/
     function getThumbPath(videoPath: string): string {
-        const file = videoPath.split('/').pop();
-
-        return `${thumbCacheFolder}/${file}.bmp`
+        return thumbnails.getThumbPath(videoPath);
     }
 
     // Get thumbnail url based on video name
     function getThumbUrl(videoPath: string): string {
-        return `file://${getThumbPath(videoPath)}`;
+        return thumbnails.getThumbUrl(videoPath);
     }
 
-    onWallpapersFolderChanged: {
-        // Reset variables
-        root._thumbGenIndex = 0;
-        
-        root.thumbRegenerate();
+    function thumbRegenerate() {
+        thumbnails.thumbRegenerate();
     }
 
-    onCurrentWallpaperChanged: {
-        if (!root.active)
-            return;
 
-        if (root.currentWallpaper != "") {
-            Logger.d("mpvpaper", "Changing current wallpaper:", root.currentWallpaper);
+    /***************************
+    * COMPONENTS
+    ***************************/
+    Mpvpaper {
+        // Contains all the mpvpaper specific functionality
+        id: mpvpaper
+        pluginApi: root.pluginApi
 
-            if(mpvProc.running) {
-                // If mpvpaper is already running
-                socket.connected = true;
-                socket.path = mpvSocket;
-                socket.write(`loadfile "${root.currentWallpaper}"\n`);
-                socket.flush();
-            } else {
-                // Start mpvpaper
-                mpvProc.command = ["sh", "-c", `mpvpaper -o "input-ipc-server=${root.mpvSocket} loop no-audio" ALL ${root.currentWallpaper}` ]
-                mpvProc.running = true;
-            }
-        } else if(mpvProc.running) {
-            Logger.d("mpvpaper", "Current wallpaper is empty, turning mpvpaper off.");
+        active: root.active
+        currentWallpaper: root.currentWallpaper
+        hardwareAcceleration: root.hardwareAcceleration
+        isMuted: root.isMuted
+        isPlaying: root.isPlaying
+        mpvSocket: root.mpvSocket
+        profile: root.profile
+        fillMode: root.fillMode
+        volume: root.volume
 
-            socket.connected = false;
-            mpvProc.running = false;
-        }
+        thumbnails: thumbnails
+        innerService: innerService
     }
 
-    onActiveChanged: {
-        if(root.active && !mpvProc.running && root.currentWallpaper != "") {
-            Logger.d("mpvpaper", "Turning mpvpaper on.");
+    Thumbnails {
+        // Contains all the thumbnail specific functionality
+        id: thumbnails
+        pluginApi: root.pluginApi
 
-            mpvProc.command = ["sh", "-c", `mpvpaper -o "input-ipc-server=${root.mpvSocket} loop no-audio" ALL ${root.currentWallpaper}` ]
-            mpvProc.running = true;
-        } else if(!root.active) {
-            Logger.d("mpvpaper", "Turning mpvpaper off.");
+        currentWallpaper: root.currentWallpaper
+        thumbCacheReady: root.thumbCacheReady
 
-            mpvProc.running = false;
-        }
+        folderModel: folderModel
     }
+
+    InnerService {
+        // Contains all the save / load functionality for this to work with noctalia
+        id: innerService
+        pluginApi: root.pluginApi
+
+        currentWallpaper: root.currentWallpaper
+        oldWallpapers: root.oldWallpapers
+
+        thumbnails: thumbnails
+    }
+
+    Automation {
+        id: automation
+        pluginApi: root.pluginApi
+
+        automation: root.automation
+        automationMode: root.automationMode
+        automationTime: root.automationTime
+
+        random: root.random
+        nextWallpaper: root.nextWallpaper
+    }
+
 
     FolderListModel {
         id: folderModel
-        folder: "file://" + root.wallpapersFolder
+        folder: root.pluginApi == null ? "" : "file://" + root.wallpapersFolder
         nameFilters: ["*.mp4", "*.avi", "*.mov"]
         showDirs: false
-    }
 
-    Process {
-        id: mpvProc
-    }
-
-    Socket {
-        id: socket
-        path: root.mpvSocket
-    }
-
-    FolderListModel {
-        id: thumbFolderModel
-        folder: "file://" + root.thumbCacheFolder
-        nameFilters: ["*.bmp"]
-        showDirs: false
-    }
-
-    Process {
-        id: thumbProc
-        onRunningChanged: {
-            if (running)
-                return;
-
-            // Try to create the thumbnails if they don't exist.
-            root.thumbGeneration();
+        onStatusChanged: {
+            if (folderModel.status == FolderListModel.Ready) {
+                // Generate all the thumbnails for the folder
+                thumbnails.thumbGeneration();
+            }
         }
     }
 
@@ -200,6 +222,7 @@ Item {
             root.clear();
         }
 
+        // Current wallpaper
         function setWallpaper(path: string) {
             root.setWallpaper(path);
         }
@@ -208,8 +231,12 @@ Item {
             return root.currentWallpaper;
         }
 
+        // Active
         function setActive(isActive: bool) {
-            root.setActive(isActive);
+            if (root.pluginApi == null) return;
+
+            root.pluginApi.pluginSettings.active = isActive;
+            root.pluginApi.saveSettings();
         }
 
         function getActive(): bool {
@@ -217,7 +244,79 @@ Item {
         }
 
         function toggleActive() {
-            root.setActive(!root.active);
+            setActive(!root.active);
+        }
+
+        // Is playing
+        function resume() {
+            if (root.pluginApi == null) return;
+
+            root.pluginApi.pluginSettings.isPlaying = true;
+            root.pluginApi.saveSettings();
+        }
+
+        function pause() {
+            if (root.pluginApi == null) return;
+
+            root.pluginApi.pluginSettings.isPlaying = false;
+            root.pluginApi.saveSettings();
+        }
+
+        function togglePlaying() {
+            if (root.pluginApi == null) return;
+
+            root.pluginApi.pluginSettings.isPlaying = !root.isPlaying;
+            root.pluginApi.saveSettings();
+        }
+
+        // Mute / unmute
+        function mute() {
+            if (root.pluginApi == null) return;
+
+            root.pluginApi.pluginSettings.isMuted = true;
+            root.pluginApi.saveSettings();
+        }
+
+        function unmute() {
+            if (root.pluginApi == null) return;
+
+            root.pluginApi.pluginSettings.isMuted = false;
+            root.pluginApi.saveSettings();
+        }
+
+        function toggleMute() {
+            if (root.pluginApi == null) return;
+
+            root.pluginApi.pluginSettings.isMuted = !root.isMuted;
+            root.pluginApi.saveSettings();
+        }
+
+        // Volume
+        function setVolume(volume: real) {
+            if (root.pluginApi == null) return;
+
+            root.pluginApi.pluginSettings.volume = volume;
+            root.pluginApi.saveSettings();
+        }
+
+        function increaseVolume() {
+            setVolume(root.volume + Settings.data.audio.volumeStep);
+        }
+
+        function decreaseVolume() {
+            setVolume(root.volume - Settings.data.audio.volumeStep);
+        }
+
+        // Hardware acceleration
+        function setHardwareAcceleration(active: bool) {
+            if (root.pluginApi == null) return;
+
+            root.pluginApi.pluginSettings.hardwareAcceleration = active;
+            root.pluginApi.saveSettings();
+        }
+
+        function toggleHardwareAcceleration() {
+            setHardwareAcceleration(!root.hardwareAcceleration);
         }
     }
 }